process_file/requirements.txt
cachetools==5.3.0 ; python_version >= "3.9" and python_version < "4.0"
certifi==2023.5.7 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.1.0 ; python_version >= "3.9" and python_version < "4.0"
dropbox==11.36.0 ; python_version >= "3.9" and python_version < "4.0"
google-api-core==2.11.0 ; python_version >= "3.9" and python_version < "4.0"
google-api-core[grpc]==2.11.0 ; python_version >= "3.9" and python_version < "4.0"
google-auth==2.18.1 ; python_version >= "3.9" and python_version < "4.0"
google-cloud-core==2.3.2 ; python_version >= "3.9" and python_version < "4.0"
google-cloud-pubsub==2.17.1 ; python_version >= "3.9" and python_version < "4.0"
google-cloud-secret-manager==2.16.1 ; python_version >= "3.9" and python_version < "4.0"
google-cloud-storage==2.9.0 ; python_version >= "3.9" and python_version < "4.0"
google-crc32c==1.5.0 ; python_version >= "3.9" and python_version < "4.0"
google-resumable-media==2.5.0 ; python_version >= "3.9" and python_version < "4.0"
googleapis-common-protos==1.59.0 ; python_version >= "3.9" and python_version < "4.0"
googleapis-common-protos[grpc]==1.59.0 ; python_version >= "3.9" and python_version < "4.0"
grpc-google-iam-v1==0.12.6 ; python_version >= "3.9" and python_version < "4.0"
grpcio-status==1.48.2 ; python_version >= "3.9" and python_version < "4.0"
grpcio==1.54.2 ; python_version >= "3.9" and python_version < "4.0"
idna==3.4 ; python_version >= "3.9" and python_version < "4.0"
ply==3.11 ; python_version >= "3.9" and python_version < "4.0"
proto-plus==1.22.2 ; python_version >= "3.9" and python_version < "4.0"
protobuf==3.20.3 ; python_version >= "3.9" and python_version < "4.0"
psycopg==3.1.18 ; python_version >= "3.9" and python_version < "4.0"
pyasn1-modules==0.3.0 ; python_version >= "3.9" and python_version < "4.0"
pyasn1==0.5.0 ; python_version >= "3.9" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.9" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.9" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.9" and python_version < "4"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
stone==3.3.1 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.11.0 ; python_version >= "3.9" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "win32"
urllib3==1.26.16 ; python_version >= "3.9" and python_version < "4.0"
